Исправление бесконечной загрузки когда элементы кончаются
