+++ backend/config.py
@@ -19,7 +19,7 @@ DB_HOST = os.getenv("DB_HOST")
 DB_PORT = os.getenv("DB_PORT")
 DB_NAME = os.getenv("DB_NAME")
 DATABASE_URL = f"postgresql+asyncpg://{DB_USER}:{DB_PASS}@{DB_HOST}:{DB_PORT}/{DB_NAME}"
-
+# DATABASE_URL = os.getenv("DATABASE_URL", "sqlite+aiosqlite:///./test.db")
 # redis connection
 REDIS_HOST = os.getenv("REDIS_HOST", "192.168.1.9")
 REDIS_PORT = int(os.getenv("REDIS_PORT", 6379))
