+++ backend/routers.py
@@ -6,23 +6,14 @@ from fastapi import Form, HTTPException
 from sqlalchemy.future import select
 from starlette.requests import Request
 from starlette.responses import HTMLResponse, RedirectResponse
-from models import Product
-from utils import check_token, add_new_item, update_current_item
+from utils import check_token, add_new_item, refresh_item
 from config import AUTH_HOST, AUTH_PORT
 from core import caching, templates
-from database import SessionLocal
+
 
 rest_router = APIRouter()
 
 
-@rest_router.get("/active_product")
-async def get_active_product():
-    async with SessionLocal() as session:
-        async with session.begin():
-            stmt = select(Product).order_by(Product.id)
-            result = await session.execute(stmt)
-            product = result.scalar_one_or_none()
-            return {"is_active": product.id, "price": product.current_price}
 
 
 @rest_router.get("/login")
@@ -31,13 +22,8 @@ async def login_get(request: Request):
     if access_token and (await check_token(access_token)):
         redirect_response = RedirectResponse(url="/", status_code=303)
         redirect_response.set_cookie(key="access_token", value=access_token)
-        active_id = caching.get("active:id").decode('utf-8')
-        if not active_id:
-            await update_current_item()
-            active_id = caching.get("active:id").decode('utf-8')
-        redirect_response.set_cookie(key="active_id", value=active_id)
         return redirect_response
-    return templates.TemplateResponse("login.html", {"request": request, "title": "WebSocket Example"})
+    return templates.TemplateResponse("login.html", {"request": request, "title": "Login"})
 
 
 @rest_router.get("/register")
@@ -67,12 +53,8 @@ async def login_post(crypto: str = Form(...), password: str = Form(...)):
                 if access_token:
                     redirect_response = RedirectResponse(url="/", status_code=303)
                     redirect_response.set_cookie(key="access_token", value=access_token)
-                    try:
-                        active_id = caching.get("active:id").decode('utf-8')
-                    except AttributeError:
-                        await update_current_item()
-                        active_id = caching.get("active:id").decode('utf-8')
-                    redirect_response.set_cookie(key="active_id", value=active_id)
+                    item = await refresh_item()
+                    redirect_response.set_cookie(key="active_id", value=item["active:id"])
                     return redirect_response
                 else:
                     raise HTTPException(status_code=401, detail="Invalid token response")
@@ -83,10 +65,14 @@ async def login_post(crypto: str = Form(...), password: str = Form(...)):
 @rest_router.get("/", response_class=HTMLResponse)
 async def read_root(request: Request):
     if request.cookies.get("access_token"):
-        item_name = caching.get("active:name").decode('utf-8')
-        img_link = caching.get("active:img_link").decode('utf-8')
-        return templates.TemplateResponse("index.html",
-                                          {"request": request, "item_name": item_name, "img_link": img_link})
+        item = await refresh_item()
+        if item:
+            item_name = item["active:name"]
+            img_link = item["active:img_link"]
+            return templates.TemplateResponse("index.html",
+                                              {"request": request, "item_name": item_name, "img_link": img_link})
+        else:
+            return RedirectResponse(url="/stopped")
     else:
         return RedirectResponse(url="/login")
 
@@ -94,6 +80,11 @@ async def read_root(request: Request):
 @rest_router.get("/new_item")
 async def new_item_get(request: Request):
     return templates.TemplateResponse("new_item.html", {"request": request, "title": "WebSocket Example"})
+
+
+@rest_router.get("/stopped")
+async def stopped(request: Request):
+    return templates.TemplateResponse("stopped.html", {"request": request})
 
 
 @rest_router.post("/new_item")
