+++ backend/static/script.js
@@ -31,7 +31,7 @@
             if (parts.length === 2) return parts.pop().split(';').shift();
     }
           const access_token = getCookie('access_token');
-          const active_id = getCookie('active_id');
+
 
           const ws = new WebSocket(`ws://${window.location.host}/ws?access_token=${access_token}`);
             function updatePrice(newPrice) {
@@ -52,25 +52,22 @@
             const data = JSON.parse(event.data);
             console.log('getMessage');
             const active_id = getCookie('active_id');
-            if (String(data.active_id) == String(active_id)) {
-                width = data.progress_bar;
-                console.log(width);
-                const audio = document.getElementById('audio-player');
-                audio.volume = 0.3;
-                audio.pause()
-                audio.play().catch(error => {
-                    console.error('Error attempting to play audio:', error);
-                });
-                document.getElementById('owner').innerText = data.address;
-                updatePrice(data.price);
-            } else {
-                console.log("update");
-                window.location.href = "/login";
-            }
+
+            width = data.progress_bar;
+            console.log(width);
+            const audio = document.getElementById('audio-player');
+            audio.volume = 0.3;
+            audio.pause()
+            audio.play().catch(error => {
+                console.error('Error attempting to play audio:', error);
+            });
+            document.getElementById('owner').innerText = data.address;
+            updatePrice(data.price);
+
         };
         ws.onclose = function(event) {
             console.log('closed');
-            window.location.href = '/login';
+            window.location.href = '/stopped';
         };
         function updateClock(timezone) {
             const clockElement = document.getElementById('clock');
