+++ backend/templates/stopped.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Auction Item</title>
+    <link rel="stylesheet" type="text/css" href="/static/styles.css">
+
+</head>
+<body>
+    <header class="header">
+        <a href="/"><div class="logo">EthB2B</div></a>
+        <nav class="nav">
+            <div class="login"><a href="/new_item">ADD</a></div>
+        </nav>
+        <div class="login"><a href="/login">UPDATE TOKEN</a></div>
+    </header>
+
+    <div class="auction-container">
+        <div class="item-info">
+            <div class="item-header">
+                <h1 class="item-name"><span id="item-name">Аукцион на данный момент не ведется</span></h1>
+
+
+            </div>
+
+        </div>
+    </div>
+</body>
+</html>
