+++ backend/utils.py
@@ -10,7 +10,8 @@ from sqlalchemy import select, not_, and_
 from database import SessionLocal
 from config import TIME_INTERVAL, S3_CONFIG, AUTH_HOST, AUTH_PORT, CRYPTO_HOST, CRYPTO_PORT
 from core import caching, logger
-from models import Product, S3Client
+from models.DB_model import Product
+from models.S3_model import S3Client
 
 logger.info("S3 connecting")
 logger.info(f"{S3_CONFIG}")
@@ -33,44 +34,64 @@ async def check_balance(address: str):
 
 
 async def accept_winner(id, time):
-    last_item = caching.get("active:id").decode('utf-8')
-    last_bid = caching.get("active:last_bid").decode('utf-8')
-    if str(id) == (last_item) and str(time) == str(last_bid):
-        async with (SessionLocal() as session):
-            async with session.begin():
-                stmt = select(Product).where(Product.id == int(id))
-                result = await session.execute(stmt)
-                product = result.scalar_one_or_none()
-                if not product.is_sold:
-                    product.is_sold = 1
-                    product.sell_counts += 1
-                    await session.commit()
-                    await update_current_item()
+    async with SessionLocal() as session:
+        async with session.begin():
+            stmt = select(Product).where(Product.id == int(id))
+            result = await session.execute(stmt)
+            product = result.scalar_one_or_none()
+            if not product.is_sold and product.last_bid < time:
+                product.is_sold = 1
+                product.sell_counts += 1
+                await session.commit()
+                await refresh_item()
 
 
-async def update_current_item(startup=False):
-    async with (SessionLocal() as session):
-        logger.info("Session connected")
+async def refresh_item(update=False, price=0, owner=''):
+    cache = caching.get("active:id")
+    if cache and not update:
+        return {
+            "active:id": cache,
+            "active:name": caching.get("active:name").decode('utf-8'),
+            "active:img_link": caching.get("active:img_link").decode('utf-8'),
+            "active:price": caching.get("active:price").decode('utf-8'),
+            "active:owner": caching.get("active:owner").decode('utf-8'),
+            "active:last_bid": caching.get("active:last_bid")
+        }
+    async with SessionLocal() as session:
+        answer = None
         async with session.begin():
             moscow_tz = pytz.timezone('Europe/Moscow')
             current_time_moscow = datetime.datetime.now(moscow_tz).replace(tzinfo=None)
+
             stmt = select(Product).where(
                 and_(
                     Product.date_of_start < current_time_moscow,
                     not_(Product.is_sold)
-                )).order_by(Product.sell_counts)
+                )
+            ).order_by(Product.sell_counts)
 
             result = await session.execute(stmt)
-
             current_item = result.scalars().first()
+            if update:
+                current_item.owner = owner
+                current_item.current_price = price
             if current_item:
-                logger.info("Cache creation started")
-                caching.set("active:id", str(current_item.id))
-                caching.set("active:name", str(current_item.name))
-                caching.set("active:img_link", str(current_item.picture_path))
-                caching.set("active:price", str(current_item.current_price))
-                caching.set("active:owner", str(current_item.owner))
-                logger.info("Cache created")
+                caching.set("active:id", str(current_item.id), ex=TIME_INTERVAL // 2)
+                caching.set("active:name", str(current_item.name), ex=TIME_INTERVAL // 2)
+                caching.set("active:img_link", str(current_item.picture_path), ex=TIME_INTERVAL // 2)
+                caching.set("active:price", str(current_item.current_price), ex=TIME_INTERVAL // 2)
+                caching.set("active:owner", str(current_item.owner), ex=TIME_INTERVAL // 2)
+                caching.set("active:last_bid", str(current_item.last_bid), ex=TIME_INTERVAL // 2)
+                answer = {
+                        "active:id": str(current_item.id),
+                        "active:name": str(current_item.name),
+                        "active:img_link": str(current_item.picture_path),
+                        "active:price": str(current_item.current_price),
+                        "active:owner": str(current_item.owner),
+                        "active:last_bid": str(current_item.last_bid)
+                    }
+                await session.commit()
+            return answer
 
 
 def run_async_function(func, *args):
@@ -90,11 +111,9 @@ async def set_price_and_owner_to_active(id, price, owner):
                 product.current_price = float(price)
                 product.owner = owner
                 await session.commit()
-                caching.set("active:price", price)
-                caching.set("active:owner", owner)
                 moscow_tz = pytz.timezone('Europe/Moscow')
                 current_time_moscow = datetime.datetime.now(moscow_tz).replace(tzinfo=None)
-                caching.set("active:last_bid", str(current_time_moscow))
+                await refresh_item(update=True, price=price, owner=owner)
                 threading.Timer(TIME_INTERVAL, run_async_function,
                                 args=(accept_winner, id, current_time_moscow)).start()
 
@@ -107,7 +126,8 @@ async def add_new_item(name: str, file: UploadFile, price: float):
                           current_price=price,
                           is_sold=0,
                           sell_counts=0,
-                          date_of_start=datetime.datetime.now(moscow_tz).replace(tzinfo=None))
+                          date_of_start=datetime.datetime.now(moscow_tz).replace(tzinfo=None),
+                          last_bid=datetime.datetime.now(moscow_tz).replace(tzinfo=None))
     async with SessionLocal() as session:
         async with session.begin():
             session.add(new_product)
